backend/requirements.txt
fastapi
uvicorn
openai
requests
feedparser
python-dotenv
pydantic
beautifulsoup4
